% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_control_parameters.R
\name{set_atsb}
\alias{set_atsb}
\title{Parameterise an ATSB strategy}
\usage{
set_atsb(parameters, timesteps, coverages)
}
\arguments{
\item{parameters}{a list of parameters to modify}

\item{timesteps}{the timesteps at which to have ATSB active}

\item{coverages}{the proportion of the human population who receive ATSB}
}
\description{
The model will switch on ATSB.
}
